Changed float to int in Enemy virtual class. Changed screen sizes in globals to be unsigned int rather than floats. Fixed Walker code so it acts appropriately. Implemented weight system for Walker, for more easily changeable fall speed manipulation.

+++ Archer/Enemy.h
@@ -19,7 +19,7 @@ protected:
 
 	sf::Vector2f m_movement;
 	sf::RectangleShape m_hitbox;
-	float m_lastPlatformCollision;
+	int m_lastPlatformCollision;
 	bool m_falling;
 
 public:
